+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+
+- `DOMNode`:
+  - `buildDOM`: added parameters `treeMap` and `setTreeMapRoot`, passed to `DOMGenerator.generate`.
+  - Added `nodesView` getter returning an unmodifiable list of children nodes.
+
 ## 3.0.0-beta.8
 
 - `DOMAttribute`: